move model out + clean up
